pyproject.toml
[project]
name = "basic-react-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.66.0",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
]
